+++ git-protocol/src/fetch/mod.rs
@@ -27,7 +27,6 @@ pub use response::Response;
 
 ///
 pub mod handshake;
-
 pub use handshake::function::handshake;
 
 #[cfg(test)]
